+++ app/packages/core/src/components/ColorModal/ColorModal.tsx
@@ -2,7 +2,6 @@ import React, { Fragment, useMemo, useRef } from "react";
 import Draggable from "react-draggable";
 import ReactDOM from "react-dom";
 import { useRecoilState, useRecoilValue } from "recoil";
-import styled from "styled-components";
 import CloseIcon from "@mui/icons-material/Close";
 import KeyboardArrowDownOutlinedIcon from "@mui/icons-material/KeyboardArrowDownOutlined";
 import KeyboardArrowUpOutlinedIcon from "@mui/icons-material/KeyboardArrowUpOutlined";
@@ -21,74 +20,14 @@ import { useOutsideClick } from "@fiftyone/state";
 import useMeasure from "react-use-measure";
 import JSONViewer from "./JSONViewer";
 import ColorFooter from "./ColorFooter";
-
-const ModalWrapper = styled.div`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  z-index: 1000000;
-  align-items: center;
-  display: flex;
-  justify-content: center;
-  background-color: ${({ theme }) => theme.neutral.softBg};
-`;
-
-type ContainerProp = {
-  height: string;
-  width: string;
-  minWidth: string;
-};
-
-const Container = styled.div<ContainerProp>`
-  background-color: ${({ theme }) => theme.background.level2};
-  border: 1px solid ${({ theme }) => theme.primary.plainBorder};
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  justify-content: start;
-  overflow: hidden;
-  box-shadow: ${({ theme }) => `0 20px 25px -20px ${theme.background.level1}`};
-  min-width: ${({ minWidth }) => minWidth};
-  width: ${({ width }) => width};
-  height: ${({ height }) => height};
-`;
-
-const DraggableContent = styled.div<ContainerProp>`
-  height: ${({ height }) => `calc(${height} - 6rem)`};
-  overflow: auto;
-`;
-
-const DraggableModalTitle = styled.div`
-  flex-direction: row;
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  height: 2.5rem;
-  background-color: ${({ theme }) => theme.background.level1};
-  padding: 2px;
-  cursor: pointer;
-  font-weight: 600;
-`;
-
-const Text = styled.div`
-  font-size: 1.1rem;
-  text-transform: uppercase;
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  cursor: pointer;
-`;
-
-const ActionDiv = styled.div`
-  position: relative;
-`;
-
-const IconDiv = styled.div`
-  margin: auto 0.25rem;
-  align-items: center;
-`;
+import {
+  ActionDiv,
+  Container,
+  DraggableContent,
+  DraggableModalTitle,
+  IconDiv,
+  ModalWrapper,
+} from "./ShareStyledDiv";
 
 type Option = {
   value: string;
